requirements.txt
Django==1.11.7
Pillow==4.3.0
dj-database-url==0.4.2
django-forms-bootstrap==3.1.0
django-geoposition==0.3.0
django-hvad==1.8.0
olefile==0.44
psycopg2==2.7.3.2
pytz==2017.3
stripe==1.75.1
